+++ redux/arc.js
@@ -5,7 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
- import { AsyncStorage } from 'react-native';
+import React from 'react'
+import { AsyncStorage, View } from 'react-native'
 
 import getTheme from '../utils/getTheme'
 
@@ -18,16 +19,54 @@ const CONSTANTS = {
     },
 }
 
+const defaultTheme = getTheme()
+
+function hexToRGB(hex, alpha) {
+    var r = parseInt(hex.slice(1, 3), 16),
+        g = parseInt(hex.slice(3, 5), 16),
+        b = parseInt(hex.slice(5, 7), 16);
+
+    if (alpha) {
+        return "rgba(" + r + ", " + g + ", " + b + ", " + alpha + ")";
+    } else {
+        return "rgb(" + r + ", " + g + ", " + b + ")";
+    }
+}
+
+function getNavigationOptions( theme ) {
+
+    const colors = theme.defaultValues.colors
+
+    return {
+        headerStyle: {
+            backgroundColor: hexToRGB(colors.headerColor, colors.headerOpacity),
+        },
+        headerTintColor: colors.headerTintColor,
+        headerTitleStyle: {
+            color: colors.headerTitle,
+        },
+        headerTransparent: colors.headerOpacity < 1,
+        }
+
+}
+
 const initialState = {
-    theme: getTheme(),
+    theme: defaultTheme,
+    navigationOptions: getNavigationOptions( defaultTheme ),
+    contentOffset: 0,
 }
 
 //actions
 
 function setTheme(theme) {
+    const navigationOptions = getNavigationOptions( theme )
+    const contentOffset = theme.defaultValues.colors.headerOpacity < 1 ? (theme.systemValues.isIOS ? theme.systemValues.statusBarHeight : 0) + theme.systemValues.navigationBarHeight : 0
     return {
         type: CONSTANTS.theme.setTheme,
         theme,
+        navigationOptions,
+        // childNavigationOptions,
+        contentOffset,
     }
 }
 
@@ -38,23 +77,45 @@ function saveTheme(themeName) {
         if (themeName)
             AsyncStorage.setItem(CONSTANTS.theme.themeKey, themeName)
         else
-            AsyncStorage.removeItem(CONSTANTS.theme.themeKey)
+            AsyncStorage.setItem(CONSTANTS.theme.themeKey, 'default')
+    //    AsyncStorage.removeItem(CONSTANTS.theme.themeKey)
     }
 }
 
-function restoreTheme( themesArray ) {
+function restoreTheme( themesArray, defTheme = null ) {
 
     return (dispatch) => {
 
         AsyncStorage.getItem(CONSTANTS.theme.themeKey)
         .then( (result) => {
             if (result) {
+                if (result === 'default') return
                 const colors = themesArray[result]
                 if (colors) dispatch( setTheme( getTheme(result, colors) ) )
+            } else {
+                if (defTheme) {
+                const colors = themesArray[defTheme]
+                if (colors) {
+                    dispatch( setTheme( getTheme(result, colors) ) )
+                    dispatch( saveTheme( defTheme ) )
+                    }
+                }
             }
         })
 
     }
+}
+
+function checkThemeChange(N, O = null, rootLevel = false) {
+    const n_theme = N.theme ? N.theme.theme||{} : {}
+
+    if (O) {
+        const o_theme = O.theme ? O.theme.theme||{} : {}
+        if (n_theme.name === o_theme.name) return
+//        if (N.theme.navigationOptions === O.theme.navigationOptions) return
+    }
+
+    N.navigation.setParams({ navigationOptions: N.theme.navigationOptions })
 }
 
 //reducer
@@ -67,6 +128,8 @@ function themeReducer(state = initialState, action) {
             return {
                 ...state,
                 theme: action.theme,
+                navigationOptions: action.navigationOptions,
+                contentOffset: action.contentOffset,
             }
         default:
             return state
@@ -75,6 +138,7 @@ function themeReducer(state = initialState, action) {
 
 export {
     CONSTANTS,
+    checkThemeChange,
     setTheme,
     saveTheme,
     restoreTheme,